Telegram-Bot/requirements.txt
aiohttp~=3.8.1
aiogram~=2.14.3
cryptography~=36.0.2
requests~=2.27.1
prometheus-client~=0.11.0
python-dotenv~=0.19.2